Renamed base classes
